feat(week2): Halfway through week 2

+++ Prac/week1/week1programs.cpp
@@ -19,7 +19,7 @@ void conversionExample() {
 }
 
 void findPosition() {
-  int salary;
+  float salary;
   cout << "Enter your salary:";
   cin >> salary;
   int category;
